requirements.txt
qrcode[pil]==8.2
pillow==11.3.0
flask==3.0.0